requirements.txt
Django==2.2.14
djangorestframework==3.11.0
python-memcached==1.59
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
